app/requirements.txt
flask==3.0.2
mysql-connector-python==9.0.0